Java 9+ support
Earlier, not-yet-commons-ssl interpreted PKCS12 keystores as 'keystores
of one key' and was not able to start on a VM equipped with
cacerts/jssecacerts in that format. Since JDK 9, PKCS12 is the default
format for cacerts/jssecacerts.
We now allow PKCS12 stores to be handled in the same way as JKS stores
are handled.